# apidocs.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Shoes API",
      "description": "List of shoes with their details",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3001",
        "description": "Local server"
      }
    ],
    "tags": [
      {
        "name": "Shoes",
        "description": "API for shoes in the system"
      }
    ],
    "paths": {
      "/shoes": {
        "get": {
          "tags": ["Shoes"],
          "summary": "Get all shoes",
          "description": "Get all shoes in the system",
          "operationId": "getShoes",
          "responses": {
            "200": {
              "description": "A list of shoes.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Shoe"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Shoes"],
          "summary": "Add a new shoe",
          "description": "Add a new shoe to the system",
          "operationId": "addShoe",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Shoe added"
            }
          }
        }
      },
      "/shoes/{id}": {
        "get": {
          "tags": ["Shoes"],
          "summary": "Get shoe by id",
          "description": "Get a shoe by its id",
          "operationId": "getShoeById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of shoe to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "A shoe.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Shoe"
                  }
                }
              }
            },
            "404": {
              "description": "Shoe not found"
            }
          }
        },
        "put": {
          "tags": ["Shoes"],
          "summary": "Update an existing shoe",
          "description": "Update an existing shoe in the system",
          "operationId": "updateShoe",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of shoe to update",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Shoe updated"
            },
            "404": {
              "description": "Shoe not found"
            }
          }
        },
        "delete": {
          "tags": ["Shoes"],
          "summary": "Delete a shoe",
          "description": "Delete a shoe in the system",
          "operationId": "deleteShoe",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of shoe to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Shoe deleted"
            },
            "404": {
              "description": "Shoe not found"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Shoe": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "brand": {
              "type": "string"
            },
            "model": {
              "type": "string"
            },
            "size": {
              "type": "integer"
            },
            "color": {
              "type": "string"
            },
            "price": {
              "type": "number",
              "format": "double"
            }
          }
        }
      }
    }
  }
  